LPS-53683 use resource URL
